requirements.txt
fastapi
uvicorn[standard]
gunicorn
torch
torchvision
onnxruntime
pillow
numpy
python-multipart
gradio